requirements.txt
sortedcontainers>= 2.0.4
numpy>= 1.10.4
pyannote.core>=4.1
cvxpy== 1.0.25
tqdm>= 4.46.0
numba>= 0.48.0
typing_extensions>= 3.7.4.3
TextGrid>=1.5
pympi-ling>=1.69
pyannote.database >= 4.0.1

